+++ SourceCode/Utility/PerformDetector/PerformResultTreeItemData.h
@@ -5,7 +5,7 @@
 class CPerformResultTreeItemData : public wxTreeItemData
 {
 public:
-    CPerformResultTreeItemData(const size_t id)
+    CPerformResultTreeItemData(const uint32_t id)
         : m_recordId(id)
     {
 
@@ -14,12 +14,12 @@ public:
     {
 
     }
-    size_t GetRecordId(){return m_recordId;}
-    void SetRecordId(const size_t& id){m_recordId = id;}
+    uint32_t GetRecordId(){return m_recordId;}
+    void SetRecordId(const uint32_t& id){m_recordId = id;}
 
 
 private:
-    size_t m_recordId;
+    uint32_t m_recordId;
 };
 
 #endif
