+++ Tool/BeyondEngineEditor/ComponentTreeItemData.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "ComponentTreeItemData.h"
 
-CComponentTreeItemData::CComponentTreeItemData(bool bIsDirectory, size_t guid)
+CComponentTreeItemData::CComponentTreeItemData(bool bIsDirectory, uint32_t guid)
     : m_bIsDirectory(bIsDirectory)
     , m_uComponentGUID(guid)
 {
@@ -16,7 +16,7 @@ bool CComponentTreeItemData::IsDirectory()
     return m_bIsDirectory;
 }
 
-size_t CComponentTreeItemData::GetGUID()
+uint32_t CComponentTreeItemData::GetGUID()
 {
     return m_uComponentGUID;
 }
